requirements.txt
# Core LLM + agent framework
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langgraph>=0.2.0

# Observability
langfuse>=2.0.0

# Vector store & embeddings
chromadb>=0.5.0
langchain-chroma>=0.1.0
sentence-transformers>=3.3.0

# Data & statistics
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
rich>=13.7.0
